Remove split from classy dataset (#274)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/274

Pull Request resolved: https://github.com/fairinternal/ClassyVision/pull/52

In writing the tutorial for the dataset, we realized that we don't need the split argument. This led to the realization that the split argument is currently in the base dataset class and it doesn't need to be there (many datasets don't need splits for specification, some like Imagenet / CIFAR do).

This diff removes it from classy dataset to simplify the API (though several down stream datasets still use a split argument).

Note, the hard part about this is that we need to make sure we change all of the downstream use-cases.

Reviewed By: mannatsingh

Differential Revision: D18777326

fbshipit-source-id: e8d2d2d26b5cfd176d8165043333ab7d0c3699c7